Alerts en Marca/Tipo

+++ thermopack/src/app/admin/pages/admin-categories/admin-categories.component.ts
@@ -4,6 +4,7 @@ import { MainService } from '../../../services/service';
 import { Brands } from '../../../interfaces/brands.interface';
 import { Types } from '../../../interfaces/types.interface';
 import { Categories } from '../../../interfaces/categories.interface';
+import { SweetAlertService } from '../../services/sweet-alert.service';
 
 @Component({
     selector: 'app-admin-categories',
@@ -19,6 +20,7 @@ export class AdminCategoriesComponent {
 
   constructor(
     private service: MainService,
+    private sweetAlertService: SweetAlertService,
   ) {}
 
   ngOnInit() {
@@ -44,23 +46,43 @@ export class AdminCategoriesComponent {
     this.service.getBrandByName(brand._id).subscribe((response) => {
       if(response._id === undefined){
         if(brand._id.trim() !== ''){
-          this.service.updateBrandByName(brand.name, brand._id).subscribe((response) => {
-            alert('Marca editada'); //TODO Alert
-          });
+          this.sweetAlertService.showConfirmationAlert(
+            'Confirmación',
+            '¿Está seguro que desea editar esta marca?',
+            () => {
+              this.service.updateBrandByName(brand.name, brand._id).subscribe((response) => {
+                this.sweetAlertService.showAlert('Éxito', 'La marca fue editada correctamente.', 'success');
+              },
+              (error) => {
+                this.sweetAlertService.showAlert('Error', 'Hubo un error al editar los datos.', 'error');
+              });
+            },
+            () => {
+              location.reload();
+            }
+          );
         }else{
-          alert('No puede registrar una marca vacía.'); //TODO Alert
+          this.sweetAlertService.showAlert('Error', 'No puede registrar una marca vacía.', 'error');
         }
       }else{
-        alert('Esta marca ya está registrada.'); //TODO Alert
+        this.sweetAlertService.showAlert('Atención', 'Esta marca ya está registrada.', 'info');
       }
     });
   }
 
   deleteBrand(brand: Brands) {
-    this.service.deleteBrandByName(brand.name).subscribe((response) => {
-      console.log(response);
-      alert('Marca eliminada');
-    })
+    this.sweetAlertService.showConfirmationAlert(
+      'Confirmación',
+      `¿Está seguro que desea eliminar la marca ${brand.name}? Asegúrese que no existan productos de esta marca.`,
+      () => {
+        this.service.deleteBrandByName(brand.name).subscribe((response) => {
+          this.sweetAlertService.showAlert('Éxito', 'La marca fue eliminada correctamente.', 'success');
+        },
+        (error) => {
+          this.sweetAlertService.showAlert('Error', 'Hubo un error al eliminar los datos.', 'error');
+        });
+      }
+    );
   }
 
   editType(type: Types) {
@@ -68,23 +90,43 @@ export class AdminCategoriesComponent {
     this.service.getTypeByName(type._id).subscribe((response) => {
       if(response._id === undefined){
         if(type._id.trim() !== ''){
-          this.service.updateTypeByName(type.name, type._id).subscribe((response) => {
-            alert('Tipo editado'); //TODO Alert
-          });
+          this.sweetAlertService.showConfirmationAlert(
+            'Confirmación',
+            '¿Está seguro que desea editar este tipo?',
+            () => {
+              this.service.updateTypeByName(type.name, type._id).subscribe((response) => {
+                this.sweetAlertService.showAlert('Éxito', 'El tipo fue editado correctamente.', 'success');
+              },
+              (error) => {
+                this.sweetAlertService.showAlert('Error', 'Hubo un error al editar los datos.', 'error');
+              });
+            },
+            () => {
+              location.reload();
+            }
+          );
         }else{
-          alert('No puede registrar un tipo vacío.'); //TODO Alert
+          this.sweetAlertService.showAlert('Error', 'No puede registrar un tipo vacío.', 'error');
         }
       }else{
-        alert('Este tipo ya está registrado.'); //TODO Alert
+        this.sweetAlertService.showAlert('Atención', 'Este tipo ya está registrado.', 'info');
       }
     })
   }
 
   deleteType(type: Types) {
-    this.service.deleteTypeByName(type.name).subscribe((response) => {
-      console.log(response);
-      alert('Tipo eliminado');
-    })
+    this.sweetAlertService.showConfirmationAlert(
+      'Confirmación',
+      `¿Está seguro que desea eliminar el tipo ${type.name}? Asegúrese que no existan productos de este tipo.`,
+      () => {
+        this.service.deleteTypeByName(type.name).subscribe((response) => {
+          this.sweetAlertService.showAlert('Éxito', 'El tipo fue eliminado correctamente.', 'success');
+        },
+        (error) => {
+          this.sweetAlertService.showAlert('Error', 'Hubo un error al eliminar los datos.', 'error');
+        });
+      }
+    );
   }
 
   editCategory(category: Categories) {
@@ -92,23 +134,46 @@ export class AdminCategoriesComponent {
     this.service.getCategoryByName(category._id).subscribe((response) => {
       if(response._id === undefined){
         if(category._id.trim() !== ''){
-          this.service.updateCategoryByName(category.name, category._id).subscribe((response) => {
-            alert('Categoría editada'); //TODO Alert
-          });
+          this.sweetAlertService.showConfirmationAlert(
+            'Confirmación',
+            '¿Está seguro que desea editar este tipo?',
+            () => {
+              this.service.updateCategoryByName(category.name, category._id).subscribe((response) => {
+                this.sweetAlertService.showAlert('Éxito', 'La categoría fue editada correctamente.', 'success');
+              },
+              (error) => {
+                this.sweetAlertService.showAlert('Error', 'Hubo un error al editar los datos.', 'error');
+              });
+            },
+            () => {
+              location.reload();
+            }
+          );
         }else{
-          alert('No puede registrar una categoría vacía.'); //TODO Alert
+          this.sweetAlertService.showAlert('Error', 'No puede registrar una categoría vacía.', 'error');
         }
       }else{
-        alert('Esta categoría ya está registrada.'); //TODO Alert
+        this.sweetAlertService.showAlert('Atención', 'Esta categoría ya está registrada.', 'info');
       }
     })
   }
 
   deleteCategory(category: Categories) {
-    this.service.deleteCategoryByName(category.name).subscribe((response) => {
-      console.log(response);
-      alert('Categoría eliminada');
-    })
+    this.sweetAlertService.showConfirmationAlert(
+      'Confirmación',
+      `¿Está seguro que desea eliminar la categoría ${category.name}? Asegúrese que no existan productos de esta categoría.`,
+      () => {
+        this.service.deleteCategoryByName(category.name).subscribe((response) => {
+          this.sweetAlertService.showAlert('Éxito', 'La categoría fue eliminada correctamente.', 'success');
+        },
+        (error) => {
+          this.sweetAlertService.showAlert('Error', 'Hubo un error al eliminar los datos.', 'error');
+        });
+      },
+      () => {
+        location.reload();
+      }
+    );
   }
 
 }
